Refactor. Use base MPC class for now
